requirements.txt
Django>=5.0.2
django-ninja>=1.0.1
python-telegram-bot==20.0
python-dotenv>=1.0.0 
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0 